cesar/OCR_Extract2json/requirements.txt
pytesseract==0.3.13
pdf2image==1.17.0
opencv-python==4.11.0.86
Pillow==11.2.1
PyYAML==6.0.2
tabulate==0.9.0
tqdm==4.67.1
torch==2.7.0

#  curl -L -o /opt/homebrew/share/tessdata/fra.traineddata https://github.com/tesseract-ocr/tessdata/raw/master/fra.traineddata
#
# export TESSDATA_PREFIX=/opt/homebrew/share/tessdata/

